fix: Support scrolling to chart popovers in scrollable containers
