+++ src/mixed-line-bar-chart/__integ__/mixed-line-bar-chart.test.ts
@@ -503,6 +503,7 @@ describe('Details popover', () => {
       await page.windowScrollTo({ top: 150 });
       const barChart = createWrapper().findBarChart();
       const barGroup = barChart.findBarGroups().get(1).toSelector();
+      await expect(page.getWindowScroll()).resolves.toEqual({ top: 150, left: 0 });
       await page.clickBarGroup(barGroup);
       await page.waitForVisible(barChart.findDetailPopover().toSelector());
       await expect(page.getText(barChart.findDetailPopover().findHeader().toSelector())).resolves.toContain('Apr 2023');
@@ -517,10 +518,28 @@ describe('Details popover', () => {
       await page.windowScrollTo({ top: 150 });
       const barChart = createWrapper().findBarChart();
       const barGroup = barChart.findBarGroups().get(1).toSelector();
-      await page.hoverElement(barGroup);
       await expect(page.getWindowScroll()).resolves.toEqual({ top: 150, left: 0 });
+      await page.hoverElement(barGroup);
       await expect(page.getText(barChart.findDetailPopover().findHeader().toSelector())).resolves.toContain('Apr 2023');
       await expect(page.getWindowScroll()).resolves.toEqual({ top: 150, left: 0 });
+    })
+  );
+
+  test(
+    'scrolls if necessary on click inside a scrollable container',
+    setupTest('#/light/bar-chart/in-modal', async page => {
+      await page.setWindowSize({ width: 360, height: 650 });
+      const wrapper = createWrapper();
+      const modalSelector = wrapper.findModal().toSelector();
+      const modalTop = (await page.getBoundingBox(wrapper.findModal().findHeader().toSelector())).top;
+      await page.elementScrollTo(modalSelector, { left: 0, top: 400 });
+      await expect(page.getElementScroll(modalSelector)).resolves.toEqual({ left: 0, top: 400 });
+      const barChart = wrapper.findBarChart();
+      const barGroup = barChart.findBarGroups().get(1).toSelector();
+      await page.clickBarGroup(barGroup);
+      await page.waitForVisible(barChart.findDetailPopover().toSelector());
+      await expect(page.getText(barChart.findDetailPopover().findHeader().toSelector())).resolves.toContain('Apr 2023');
+      await expect(page.getElementScroll(modalSelector)).resolves.toEqual({ left: 0, top: modalTop });
     })
   );
